ml/requirements.txt
pandas
numpy
sentence_transformers
transformers
torch
datasets
flask
flask_cors
openai
python-dotenv